move copy/pasted read_vpd function into new drv-packrat-vpd-loader crate
