+++ drv/gimlet-seq-server/Cargo.toml
@@ -9,14 +9,13 @@ drv-gimlet-seq-api = { path = "../gimlet-seq-api" }
 drv-i2c-api = { path = "../i2c-api" }
 drv-i2c-devices = { path = "../i2c-devices" }
 drv-ice40-spi-program = { path = "../ice40-spi-program" }
-drv-local-vpd = { path = "../local-vpd" }
+drv-packrat-vpd-loader = { path = "../packrat-vpd-loader" }
 drv-spi-api = { path = "../spi-api" }
 drv-stm32h7-spi = { path = "../stm32h7-spi" }
 drv-stm32xx-sys-api = { path = "../stm32xx-sys-api" }
 gnarle = { path = "../../lib/gnarle" }
 ringbuf = { path = "../../lib/ringbuf" }
 task-jefe-api = { path = "../../task/jefe-api" }
-task-packrat-api = { path = "../../task/packrat-api" }
 userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
 
 byteorder = { workspace = true }
