+++ drv/packrat-vpd-loader/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "drv-packrat-vpd-loader"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+drv-local-vpd.path = "../local-vpd"
+ringbuf.path = "../../lib/ringbuf"
+task-packrat-api.path = "../../task/packrat-api"
+userlib.path = "../../sys/userlib"
