+++ drv/psc-seq-server/Cargo.toml
@@ -4,11 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-drv-local-vpd = { path = "../local-vpd" }
+drv-packrat-vpd-loader = { path = "../packrat-vpd-loader" }
 drv-psc-seq-api = { path = "../psc-seq-api" }
-ringbuf = { path = "../../lib/ringbuf" }
 task-jefe-api = { path = "../../task/jefe-api" }
-task-packrat-api = { path = "../../task/packrat-api" }
 userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
 
 # This section is here to discourage RLS/rust-analyzer from doing test builds,
