+++ drv/sidecar-seq-server/Cargo.toml
@@ -14,12 +14,11 @@ zerocopy = { workspace = true }
 drv-fpga-api = { path = "../fpga-api", features = ["auxflash"] }
 drv-i2c-api = { path = "../i2c-api" }
 drv-i2c-devices = { path = "../i2c-devices" }
-drv-local-vpd = { path = "../local-vpd" }
+drv-packrat-vpd-loader = { path = "../packrat-vpd-loader" }
 drv-sidecar-front-io = { path = "../sidecar-front-io", features = ["controller", "phy_smi"] }
 drv-sidecar-mainboard-controller = { path = "../sidecar-mainboard-controller", features = ["bitstream"] }
 drv-sidecar-seq-api = { path = "../sidecar-seq-api" }
 ringbuf = { path = "../../lib/ringbuf" }
-task-packrat-api = { path = "../../task/packrat-api" }
 userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
 
 [features]
